experiment: make any arbitrary element a trigger for opening Coloris picker
Coloris only officially supports binding to input elements. In this commit, I've figured out a way to customize Coloris so that any element, not just inputs, can act as the trigger for opening the picker, and also the color preview label is hidden (by overriding the display property in CSS).